Rename C++ sources to .cc and add C ABI entry point
