+++ .github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
       - uses: actions/checkout@v3
       - run: |
           sudo apt-get update && sudo apt-get install -y clang-format
-          find src include tests -iname '*.cpp*' | xargs clang-format --dry-run --Werror
+          find src include tests -iname '*.cc' | xargs clang-format --dry-run --Werror
 
   static-analysis:
     runs-on: ubuntu-latest
@@ -20,7 +20,7 @@ jobs:
       - uses: actions/checkout@v3
       - run: |
           sudo apt-get update && sudo apt-get install -y clang-tidy
-          find src include -iname '*.cpp*' | xargs clang-tidy -quiet
+          find src include -iname '*.cc' | xargs clang-tidy -quiet
 
   build-test:
     runs-on: ubuntu-latest
