+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_library(ourlang INTERFACE)
 target_include_directories(ourlang INTERFACE ${CMAKE_CURRENT_LIST_DIR}/../include)
 
-add_executable(ouro_lang main.cpp)
+add_executable(ouro_lang main.cc repl_binding.cc)
 target_link_libraries(ouro_lang PRIVATE ourlang)
 
 set_target_properties(ouro_lang PROPERTIES CXX_STANDARD 23 CXX_STANDARD_REQUIRED YES)
