+++ src/core.cc
