+++ src/core.impl.cc
