+++ src/main.cc
@@ -1,4 +1,4 @@
-#include "ourolang/repl.hpp"
+#include "ourolang/repl.h"
 
 int main() {
     ouro::repl();
