+++ src/repl_binding.cc
@@ -1,0 +1,5 @@
+#include "ourolang/repl.h"
+
+void ouro_repl() {
+    ouro::repl();
+}
