+++ tests/lexer_tests.cc
